Split client construction tests out of client spec
